+++ __test__/server/api/campaign/campaign.test.js
@@ -1,17 +1,16 @@
-/* eslint-disable no-unused-expressions, consistent-return */
 import gql from "graphql-tag";
-import { r } from "../../../src/server/models/";
-import { dataQuery as TexterTodoListQuery } from "../../../src/containers/TexterTodoList";
-import { dataQuery as TexterTodoQuery } from "../../../src/containers/TexterTodo";
-import { campaignDataQuery as AdminCampaignEditQuery } from "../../../src/containers/AdminCampaignEdit";
-import { campaignsQuery } from "../../../src/containers/PaginatedCampaignsRetriever";
+import { r } from "../../../../src/server/models";
+import { dataQuery as TexterTodoListQuery } from "../../../../src/containers/TexterTodoList";
+import { dataQuery as TexterTodoQuery } from "../../../../src/containers/TexterTodo";
+import { campaignDataQuery as AdminCampaignEditQuery } from "../../../../src/containers/AdminCampaignEdit";
+import { campaignsQuery } from "../../../../src/containers/PaginatedCampaignsRetriever";
 
 import {
   bulkReassignCampaignContactsMutation,
   reassignCampaignContactsMutation
-} from "../../../src/containers/AdminIncomingMessageList";
+} from "../../../../src/containers/AdminIncomingMessageList";
 
-import { makeTree } from "../../../src/lib";
+import { makeTree } from "../../../../src/lib";
 
 import {
   setupTest,
@@ -32,9 +31,8 @@ import {
   getCampaignContact,
   sendMessage,
   bulkSendMessages,
-  runGql,
-  createStartedCampaign
-} from "../../test_helpers";
+  runGql
+} from "../../../test_helpers";
 
 let testAdminUser;
 let testInvite;
@@ -94,7 +92,7 @@ afterEach(async () => {
 }, global.DATABASE_SETUP_TEARDOWN_TIMEOUT);
 
 it("allow supervolunteer to retrieve campaign data", async () => {
-  let campaignDataResults = await runComponentGql(
+  const campaignDataResults = await runComponentGql(
     AdminCampaignEditQuery,
     { campaignId: testCampaign.id },
     testSuperVolunteerUser
@@ -314,8 +312,8 @@ it("save campaign interaction steps, edit it, make sure the last value is set", 
   expect(copiedCampaign1.data.copyCampaign.id).not.toEqual(testCampaign.id);
 
   const prevCampaignIsteps = campaignDataResults.data.campaign.interactionSteps;
-  const compareToLater = async (campaignId, prevCampaignIsteps) => {
-    const campaignDataResults = await runComponentGql(
+  const compareToLater = async (campaignId, innerPrevCampaignIsteps) => {
+    campaignDataResults = await runComponentGql(
       AdminCampaignEditQuery,
       { campaignId },
       testAdminUser
@@ -334,10 +332,10 @@ it("save campaign interaction steps, edit it, make sure the last value is set", 
     // make sure the copied steps are new ones
     expect(
       Number(campaignDataResults.data.campaign.interactionSteps[0].id)
-    ).toBeGreaterThan(Number(prevCampaignIsteps[1].id));
+    ).toBeGreaterThan(Number(innerPrevCampaignIsteps[1].id));
     expect(
       Number(campaignDataResults.data.campaign.interactionSteps[1].id)
-    ).toBeGreaterThan(Number(prevCampaignIsteps[1].id));
+    ).toBeGreaterThan(Number(innerPrevCampaignIsteps[1].id));
     return campaignDataResults;
   };
   const campaign1Results = await compareToLater(
@@ -422,18 +420,14 @@ describe("Caching", async () => {
       await startCampaign(testAdminUser, testCampaign);
 
       queryLog = [];
-      console.log("STARTING TEXTING");
+      console.log("STARTING TEXTING"); // eslint-disable-line no-console
       for (let i = 0; i < 5; i++) {
-        const messageResult = await sendMessage(
-          testContacts[i].id,
-          testTexterUser,
-          {
-            userId: testTexterUser.id,
-            contactNumber: testContacts[i].cell,
-            text: "test text",
-            assignmentId
-          }
-        );
+        await sendMessage(testContacts[i].id, testTexterUser, {
+          userId: testTexterUser.id,
+          contactNumber: testContacts[i].cell,
+          text: "test text",
+          assignmentId
+        });
       }
       // should only have done updates and inserts
       expect(
@@ -491,16 +485,12 @@ describe("Reassignments", async () => {
     );
     // send some texts
     for (let i = 0; i < 5; i++) {
-      const messageResult = await sendMessage(
-        testContacts[i].id,
-        testTexterUser,
-        {
-          userId: testTexterUser.id,
-          contactNumber: testContacts[i].cell,
-          text: "test text",
-          assignmentId
-        }
-      );
+      await sendMessage(testContacts[i].id, testTexterUser, {
+        userId: testTexterUser.id,
+        contactNumber: testContacts[i].cell,
+        text: "test text",
+        assignmentId
+      });
     }
     // TEXTER 1 (95 needsMessage, 5 needsResponse)
     texterCampaignDataResults = await runComponentGql(
@@ -578,13 +568,13 @@ describe("Reassignments", async () => {
     expect(texterCampaignDataResults.data.assignment.allContactsCount).toEqual(
       20
     );
-    let assignmentContacts2 =
+    const assignmentContacts2 =
       texterCampaignDataResults.data.assignment.contacts;
     for (let i = 0; i < 5; i++) {
       const contact = testContacts.filter(
-        c => assignmentContacts2[i].id == c.id
+        c => assignmentContacts2[i].id === c.id.toString()
       )[0];
-      const messageResult = await sendMessage(contact.id, testTexterUser2, {
+      await sendMessage(contact.id, testTexterUser2, {
         userId: testTexterUser2.id,
         contactNumber: contact.cell,
         text: "test text autorespond",
@@ -629,11 +619,15 @@ describe("Reassignments", async () => {
     expect(texterCampaignDataResults.data.assignment.allContactsCount).toEqual(
       20
     );
+    const makeFilterFunction = contactToMatch => contactToTest =>
+      contactToMatch.id === contactToTest.id.toString();
     for (let i = 0; i < 3; i++) {
       const contact = testContacts.filter(
-        c => texterCampaignDataResults.data.assignment.contacts[i].id == c.id
+        makeFilterFunction(
+          texterCampaignDataResults.data.assignment.contacts[i]
+        )
       )[0];
-      const messageResult = await sendMessage(contact.id, testTexterUser2, {
+      await sendMessage(contact.id, testTexterUser2, {
         userId: testTexterUser2.id,
         contactNumber: contact.cell,
         text: "keep talking",
@@ -1012,7 +1006,6 @@ describe("Bulk Send", async () => {
 
 describe("campaigns query", async () => {
   let testCampaign2;
-  let testCampaign3;
 
   const cursor = {
     offset: 0,
@@ -1021,7 +1014,7 @@ describe("campaigns query", async () => {
 
   beforeEach(async () => {
     testCampaign2 = await createCampaign(testAdminUser, testOrganization);
-    testCampaign3 = await createCampaign(testAdminUser, testOrganization);
+    await createCampaign(testAdminUser, testOrganization);
   });
 
   it("correctly filters by a single campaign id", async () => {
@@ -1204,236 +1197,5 @@ describe("all interaction steps fields travel round trip", () => {
         ]);
       });
     });
-  });
-});
-
-describe("mutations.updateQuestionResponses", () => {
-  let adminUser;
-  let campaign;
-  let texterUser;
-  let contacts;
-  let assignment;
-  let interactionSteps;
-  let returnedInteractionSteps;
-  let colorInteractionSteps;
-  let redInteractionStep;
-  let shadesOfRedInteractionSteps;
-
-  beforeEach(async () => {
-    const startedCampaign = await createStartedCampaign();
-
-    ({
-      testAdminUser: adminUser,
-      testCampaign: campaign,
-      testTexterUser: texterUser,
-      testContacts: contacts,
-      assignment
-    } = startedCampaign);
-
-    interactionSteps = [
-      {
-        id: "new_1",
-        questionText: "What is your favorite color",
-        script: "Hello {firstName}. Let's talk about your favorite color.",
-        answerOption: "",
-        answerActions: "",
-        answerActionsData: "",
-        parentInteractionId: null,
-        isDeleted: false,
-        interactionSteps: [
-          {
-            id: "new_2",
-            questionText: "What is your favorite shade of red?",
-            script: "Red is an awesome color, {firstName}!",
-            answerOption: "Red",
-            answerActions: "",
-            answerActionsData: "",
-            parentInteractionId: "new_1",
-            isDeleted: false,
-            interactionSteps: [
-              {
-                id: "new_21",
-                questionText: "",
-                script: "Crimson is a rad shade of red, {firstName}",
-                answerOption: "Crimson",
-                answerActions: "",
-                answerActionsData: "",
-                parentInteractionId: "new_2",
-                isDeleted: false,
-                interactionSteps: []
-              },
-              {
-                id: "new_22",
-                questionText: "",
-                script: "Firebrick is a rad shade of red, {firstName}",
-                answerOption: "Firebrick",
-                answerActions: "",
-                answerActionsData: "",
-                parentInteractionId: "new_2",
-                isDeleted: false,
-                interactionSteps: []
-              }
-            ]
-          },
-          {
-            id: "new_3",
-            questionText: "",
-            script: "Purple is an awesome color, {firstName}!",
-            answerOption: "Purple",
-            answerActions: "",
-            answerActionsData: "",
-            parentInteractionId: "new_1",
-            isDeleted: false,
-            interactionSteps: []
-          }
-        ]
-      }
-    ];
-
-    returnedInteractionSteps = (
-      await createScript(adminUser, campaign, {
-        interactionSteps: interactionSteps[0],
-        campaignGqlFragment: `
-        interactionSteps {
-          id
-          questionText
-          script
-          answerOption
-          answerActions
-          answerActionsData
-          parentInteractionId
-          isDeleted
-        }
-      `
-      })
-    ).data.editCampaign.interactionSteps;
-
-    colorInteractionSteps = returnedInteractionSteps.filter(
-      interactionStep =>
-        interactionStep.parentInteractionId === returnedInteractionSteps[0].id
-    );
-
-    redInteractionStep = colorInteractionSteps.find(
-      colorInteractionStep => colorInteractionStep.answerOption === "Red"
-    );
-
-    shadesOfRedInteractionSteps = returnedInteractionSteps.filter(
-      interactionStep =>
-        interactionStep.parentInteractionId === redInteractionStep.id
-    );
-
-    const promises = contacts.slice(0, 2).map(contact => {
-      return sendMessage(contact.id, texterUser, {
-        text: returnedInteractionSteps[0].script,
-        contactNumber: contact.cell,
-        assignmentId: assignment.id,
-        userId: texterUser.id.toString()
-      });
-    });
-
-    const sendResults = await Promise.all(promises);
-
-    expect(sendResults).toHaveLength(2);
-    expect(sendResults[0].data.sendMessage.messageStatus).toEqual("messaged");
-    expect(sendResults[1].data.sendMessage.messageStatus).toEqual("messaged");
-  });
-
-  it("records answers for a contact", async () => {
-    const updateQuestionResponseGql = `
-      mutation updateQuestionResponses($qr: [QuestionResponseInput], $ccid: String!) {
-        updateQuestionResponses(questionResponses: $qr, campaignContactId: $ccid) {
-          id
-          messageStatus
-          questionResponseValues {
-            interactionStepId
-            value
-          }
-        }
-      }
-    `;
-
-    const variables = {
-      ccid: contacts[0].id,
-      qr: [
-        {
-          campaignContactId: contacts[0].id,
-          interactionStepId: returnedInteractionSteps[0].id,
-          value: colorInteractionSteps[0].answerOption
-        },
-        {
-          campaignContactId: contacts[0].id.toString(),
-          interactionStepId: redInteractionStep.id,
-          value: shadesOfRedInteractionSteps[0].answerOption
-        }
-      ]
-    };
-
-    const updateQuestionResponseResult = await runGql(
-      updateQuestionResponseGql,
-      variables,
-      texterUser
-    );
-
-    expect(updateQuestionResponseResult.data.updateQuestionResponses).toEqual({
-      id: contacts[0].id.toString(),
-      messageStatus: "messaged",
-      questionResponseValues: [
-        {
-          interactionStepId: Number(returnedInteractionSteps[0].id),
-          value: colorInteractionSteps[0].answerOption
-        },
-        {
-          interactionStepId: Number(colorInteractionSteps[0].id),
-          value: shadesOfRedInteractionSteps[0].answerOption
-        }
-      ]
-    });
-
-    const databaseQuery = `
-      SELECT
-        child.answer_option,
-        child.id as child_id,
-        child.parent_interaction_id,
-        interaction_step.campaign_id,
-        interaction_step.question,
-        interaction_step.script,
-        child.answer_actions,
-        question_response.value
-      FROM
-        question_response
-      JOIN
-        interaction_step ON interaction_step.id = question_response.interaction_step_id
-      JOIN
-        interaction_step as child ON child.parent_interaction_id = interaction_step.id and question_response.value = child.answer_option
-      WHERE
-        question_response.campaign_contact_id = ?;
-      `;
-
-    const { rows: databaseQueryResults } = await r.knex.raw(databaseQuery, [
-      contacts[0].id
-    ]);
-
-    expect(databaseQueryResults).toEqual([
-      {
-        answer_option: "Red",
-        child_id: 2,
-        parent_interaction_id: 1,
-        campaign_id: 2,
-        question: "What is your favorite color",
-        script: "Hello {firstName}. Let's talk about your favorite color.",
-        answer_actions: "",
-        value: "Red"
-      },
-      {
-        answer_option: "Crimson",
-        child_id: 3,
-        parent_interaction_id: 2,
-        campaign_id: 2,
-        question: "What is your favorite shade of red?",
-        script: "Red is an awesome color, {firstName}!",
-        answer_actions: "",
-        value: "Crimson"
-      }
-    ]);
   });
 });
